Completion of Basic Authentication (#26)
* Completed basic auth.

Completed basic auth with custom user form. Performed troubleshooting and fixed issue with email being a "required field" to use it as the username field.

* Created admin user, test user, login auth broken

Created admin user and test user to log in with. Registration works and automatically logs user in if completed successfully. Login does not work, authentication causing issues.

* Fixed authentication backends

Fixed authentication backends to allow logging in through the admin portal.

* Fixed authenticate.