Tristan toothaker (#41)
* added spotify client credentials flow

Added the Spotify client credentials flow in order to pull the list of genres from Spotify to use with the user preferences page

* Small tweaks and added correct version of channels

Figured out that the version of channels needed to be brought down to 3.0.5 in order for it to work. Also some small changes here and there.

* Basic Chat functions

Got the basic chat room functions to work, no username/room name displayed

* updated frontend

Got most of the front end messaging done, with a new design for the chat messages, and it auto scrolls to the bottom in order to keep the chat neat without extending the page.

* chatroom creation beta

This is the start of the user generation of chat room objects, it doesn't work at the moment but i wanted to commit the changes that were made

* finished up creating page

Completed the listening room creation page, added the link to the front end, created the global listening session and placed a link on the home page, small bug fixes related to the front end

* fixed l_room_create method

Fixed the issue with the chat room creation method, as well as added in error checking for if the room already exists.